# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /auth/register:
    post:
      description: Creates a new user account. By default, email verification is required.
      operationId: AuthController_register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponseDto'
        '400':
          description: Invalid input or email already exists
      summary: Register a new user
      tags:
        - auth
  /auth/login:
    post:
      description: Authenticates a user and returns access and refresh tokens
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
        '401':
          description: Invalid credentials
      summary: Login user
      tags:
        - auth
  /auth/refresh:
    post:
      description: Generates a new access token using a valid refresh token
      operationId: AuthController_refresh
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenDto'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
        '401':
          description: Invalid or expired refresh token
      summary: Refresh access token
      tags:
        - auth
  /auth/logout:
    post:
      description: Invalidates the refresh token and logs out the user
      operationId: AuthController_logout
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenDto'
      responses:
        '204':
          description: Logout successful
        '401':
          description: Invalid refresh token
      summary: Logout user
      tags:
        - auth
  /auth/verify-email:
    get:
      description: Verifies user email using the token sent via email
      operationId: AuthController_verifyEmail
      parameters:
        - name: token
          required: true
          in: query
          description: Email verification token received via email
          schema:
            example: abc123def456
            type: string
      responses:
        '200':
          description: Email verified successfully
        '400':
          description: Invalid or expired token
      summary: Verify email address
      tags:
        - auth
  /auth/resend-verification:
    post:
      description: Sends a new email verification link to the authenticated user
      operationId: AuthController_resendVerification
      parameters: []
      responses:
        '200':
          description: Verification email sent
        '401':
          description: Unauthorized
      security:
        - bearer: []
      summary: Resend verification email
      tags:
        - auth
  /auth/request-password-reset:
    post:
      description: Sends a password reset link to the provided email address if it exists
      operationId: AuthController_requestPasswordReset
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestPasswordResetDto'
      responses:
        '200':
          description: Password reset request processed
      summary: Request password reset
      tags:
        - auth
  /auth/reset-password:
    post:
      description: Resets user password using the token received via email
      operationId: AuthController_resetPassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordDto'
      responses:
        '200':
          description: Password reset successfully
        '400':
          description: Invalid or expired token
      summary: Reset password
      tags:
        - auth
...
info:
  title: Todo App API
  description: >-
    REST API for Todo Application with authentication and role-based access
    control
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    RegisterDto:
      type: object
      properties:
        email:
          type: string
          description: User email address
          example: newuser@example.com
        password:
          type: string
          description: User password (minimum 8 characters)
          example: securePassword123
          minLength: 8
          maxLength: 128
        fullName:
          type: string
          description: User full name
          example: John Doe
          minLength: 1
          maxLength: 255
        autoverify:
          type: boolean
          description: Auto-verify email (for testing purposes)
          example: false
          default: false
        role:
          type: string
          description: User role
          enum:
            - guest
            - admin
            - sysadmin
          example: guest
          default: guest
      required:
        - email
        - password
        - fullName
    RegisteredUserInfo:
      type: object
      properties:
        id:
          type: string
          description: User ID
          example: uuid-123
        email:
          type: string
          description: User email
          example: newuser@example.com
        fullName:
          type: string
          description: User full name
          example: John Doe
        role:
          type: string
          description: User role
          enum:
            - guest
            - admin
            - sysadmin
          example: guest
        emailVerified:
          type: boolean
          description: Email verification status
          example: false
      required:
        - id
        - email
        - fullName
        - role
        - emailVerified
    RegisterResponseDto:
      type: object
      properties:
        user:
          description: Registered user information
          allOf:
            - $ref: '#/components/schemas/RegisteredUserInfo'
      required:
        - user
    LoginDto:
      type: object
      properties:
        email:
          type: string
          description: User email address
          example: user@example.com
        password:
          type: string
          description: User password
          example: securePassword123
      required:
        - email
        - password
    UserInfo:
      type: object
      properties:
        id:
          type: string
          description: User ID
          example: uuid-123
        email:
          type: string
          description: User email (deprecated, use specific identity emails)
          example: user@example.com
        fullName:
          type: string
          description: User full name
          example: John Doe
        role:
          type: string
          description: User role
          enum:
            - guest
            - admin
            - sysadmin
          example: guest
        emailVerified:
          type: boolean
          description: Email verification status
          example: true
        localUsername:
          type: string
          description: Local username if local identity exists
          example: john.doe
        googleEmail:
          type: string
          description: Google email if Google identity is linked
          example: user@gmail.com
        msEmail:
          type: string
          description: Microsoft email if Microsoft identity is linked
          example: user@outlook.com
      required:
        - id
        - email
        - fullName
        - role
        - emailVerified
    AuthResponseDto:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          description: JWT refresh token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          description: User information
          allOf:
            - $ref: '#/components/schemas/UserInfo'
      required:
        - accessToken
        - refreshToken
        - user
    RefreshTokenDto:
      type: object
      properties:
        refreshToken:
          type: string
          description: Refresh token received from login or previous refresh
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      required:
        - refreshToken
    RequestPasswordResetDto:
      type: object
      properties:
        email:
          type: string
          description: Email address to send password reset link
          example: user@example.com
      required:
        - email
    ResetPasswordDto:
      type: object
      properties:
        token:
          type: string
          description: Password reset token received via email
          example: abc123def456
        newPassword:
          type: string
          description: New password (minimum 8 characters)
          example: newSecurePassword123
          minLength: 8
      required:
        - token
        - newPassword
...